test(runtime): share recording writer ingress double
Finding: AV3-M10
