+++ crates/atm-runtime-test-support/Cargo.toml
@@ -15,5 +15,7 @@ name = "atm_runtime_test_support"
 
 [dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.4.6", features = ["test-utils"] }
-atm-runtime = { path = "../atm-runtime", version = "1.4.6", features = ["test-utils"] }
+atm-runtime = { path = "../atm-runtime", version = "1.4.6" }
+atm-storage = { path = "../atm-storage", version = "1.4.6" }
 atm-storage-rusqlite = { path = "../atm-storage-rusqlite", version = "1.4.6", features = ["test-support"] }
+async-trait.workspace = true
