+++ crates/atm-runtime/Cargo.toml
@@ -12,9 +12,6 @@ homepage.workspace = true
 [lib]
 name = "atm_runtime"
 
-[features]
-test-utils = []
-
 [dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.4.6" }
 atm-storage = { path = "../atm-storage", version = "1.4.6" }
@@ -23,5 +20,6 @@ tokio.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
+atm-runtime-test-support = { path = "../atm-runtime-test-support" }
 atm-storage = { path = "../atm-storage", version = "1.4.6", features = ["test-utils"] }
 serde_json.workspace = true
