+++ crates/atm-runtime/src/mailbox_runtime.rs
@@ -409,136 +409,6 @@ pub struct StorageAsyncMailboxRuntime {
     state_handoff: StateHandoffLifecycle,
 }
 
-/// Test-only writer ingress instrumentation for router behavior gates.
-///
-/// It remains in this crate because `AsyncMessageStore` extends the sealed
-/// storage contract; HTTP-runtime tests can observe handoff effects without
-/// creating an unauthorized writer implementation at that boundary.
-#[cfg(feature = "test-utils")]
-pub mod test_utils {
-    use super::*;
-    use atm_storage::MessageStore;
-
-    #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-    pub enum RecordedWriterOutcome {
-        Applied,
-        Rejected,
-    }
-
-    #[derive(Clone, Debug, Eq, PartialEq)]
-    pub struct RecordedWriterSubmission {
-        pub message_ids: Vec<MessageKey>,
-        pub outcome: RecordedWriterOutcome,
-    }
-
-    #[derive(Clone, Default)]
-    pub struct RecordingWriterIngress {
-        inner: Arc<RecordingWriterIngressInner>,
-    }
-
-    #[derive(Default)]
-    struct RecordingWriterIngressInner {
-        reject: std::sync::atomic::AtomicBool,
-        submissions: Mutex<Vec<RecordedWriterSubmission>>,
-    }
-
-    impl RecordingWriterIngress {
-        #[must_use]
-        pub fn submissions(&self) -> Vec<RecordedWriterSubmission> {
-            self.inner
-                .submissions
-                .lock()
-                .expect("recording writer ingress mutex")
-                .clone()
-        }
-
-        pub fn set_rejecting(&self, rejecting: bool) {
-            self.inner
-                .reject
-                .store(rejecting, std::sync::atomic::Ordering::Release);
-        }
-    }
-
-    pub fn mailbox_runtime_with_recording_ingress(
-        reader: Arc<dyn AsyncMailboxReader + Send + Sync>,
-        config: HandoffConfig,
-    ) -> Result<(StorageAsyncMailboxRuntime, RecordingWriterIngress), AtmError> {
-        let ingress = RecordingWriterIngress::default();
-        let runtime = StorageAsyncMailboxRuntime::new(
-            reader,
-            Arc::new(RecordingWriter {
-                ingress: ingress.clone(),
-            }),
-        )
-        .with_state_handoff(config)?;
-        Ok((runtime, ingress))
-    }
-
-    struct RecordingWriter {
-        ingress: RecordingWriterIngress,
-    }
-
-    impl atm_storage::contract::sealed::Sealed for RecordingWriter {}
-
-    impl MessageStore for RecordingWriter {
-        fn save_message(&self, _message: &Message) -> Result<(), AtmError> {
-            unreachable!("behavior recorder receives only read-display transitions")
-        }
-
-        fn save_messages_atomically(&self, _messages: &[Message]) -> Result<(), AtmError> {
-            unreachable!("behavior recorder receives only read-display transitions")
-        }
-
-        fn load_message(&self, _key: &MessageKey) -> Result<Option<Message>, AtmError> {
-            unreachable!("behavior recorder receives only read-display transitions")
-        }
-
-        fn list_messages(&self, _query: &MessageQuery) -> Result<Vec<Message>, AtmError> {
-            unreachable!("behavior recorder receives only read-display transitions")
-        }
-
-        fn delete_message(&self, _key: &MessageKey) -> Result<(), AtmError> {
-            unreachable!("behavior recorder receives only read-display transitions")
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl AsyncMessageStore for RecordingWriter {
-        async fn apply_read_display_state_async(
-            &self,
-            _scope: MailboxScope,
-            message_ids: Vec<MessageKey>,
-            _seen_watermark: Option<IsoTimestamp>,
-        ) -> Result<(), AtmError> {
-            let rejecting = self
-                .ingress
-                .inner
-                .reject
-                .load(std::sync::atomic::Ordering::Acquire);
-            self.ingress
-                .inner
-                .submissions
-                .lock()
-                .expect("recording writer ingress mutex")
-                .push(RecordedWriterSubmission {
-                    message_ids,
-                    outcome: if rejecting {
-                        RecordedWriterOutcome::Rejected
-                    } else {
-                        RecordedWriterOutcome::Applied
-                    },
-                });
-            if rejecting {
-                Err(AtmError::daemon_unavailable(
-                    "test recording writer ingress rejected read-state handoff",
-                ))
-            } else {
-                Ok(())
-            }
-        }
-    }
-}
-
 /// Before startup composition temporarily holds the writer ingress needed to
 /// create the supervisor. After startup mailbox reads have no direct writer
 /// handle at all.
@@ -887,6 +757,7 @@ fn read_deadline(deadline: RequestDeadline) -> Result<ReadDeadline, AtmError> {
 
 #[cfg(test)]
 mod tests {
+    use atm_runtime_test_support::RecordingWriter;
     use std::path::PathBuf;
     use std::sync::Arc;
     use std::time::Duration;
@@ -898,8 +769,8 @@ mod tests {
     use atm_core::types::ReadSelection;
     use atm_storage::testing::InMemoryMailboxReader;
     use atm_storage::{
-        AgentName, AsyncMessageStore, AtmError, AtmErrorCode, IsoTimestamp, MailboxScope, Message,
-        MessageEnvelope, MessageKey, MessageQuery, MessageStore, ReadLaneError, TeamName,
+        AgentName, AtmError, AtmErrorCode, IsoTimestamp, Message, MessageEnvelope, ReadLaneError,
+        TeamName,
     };
 
     use super::{
@@ -1392,97 +1263,6 @@ mod tests {
             panic!("composition outside Tokio must not admit reads without a supervisor");
         };
         assert!(error.detail().contains("Tokio runtime"));
-    }
-
-    struct RecordingWriter {
-        failures_remaining: std::sync::atomic::AtomicUsize,
-        attempts: std::sync::atomic::AtomicUsize,
-        applied_transitions: std::sync::Mutex<Vec<Vec<MessageKey>>>,
-    }
-
-    impl Default for RecordingWriter {
-        fn default() -> Self {
-            Self::with_failures(0)
-        }
-    }
-
-    impl RecordingWriter {
-        fn with_failures(failures: usize) -> Self {
-            Self {
-                failures_remaining: std::sync::atomic::AtomicUsize::new(failures),
-                attempts: std::sync::atomic::AtomicUsize::new(0),
-                applied_transitions: std::sync::Mutex::new(Vec::new()),
-            }
-        }
-
-        fn attempts(&self) -> usize {
-            self.attempts.load(std::sync::atomic::Ordering::Acquire)
-        }
-
-        fn set_failures(&self, failures: usize) {
-            self.failures_remaining
-                .store(failures, std::sync::atomic::Ordering::Release);
-        }
-
-        fn applied(&self) -> Vec<Vec<MessageKey>> {
-            self.applied_transitions
-                .lock()
-                .expect("recording writer mutex")
-                .clone()
-        }
-    }
-
-    impl atm_storage::contract::sealed::Sealed for RecordingWriter {}
-
-    impl MessageStore for RecordingWriter {
-        fn save_message(&self, _message: &Message) -> Result<(), AtmError> {
-            unreachable!("handoff test writer only receives display transitions")
-        }
-
-        fn save_messages_atomically(&self, _messages: &[Message]) -> Result<(), AtmError> {
-            unreachable!("handoff test writer only receives display transitions")
-        }
-
-        fn load_message(&self, _key: &MessageKey) -> Result<Option<Message>, AtmError> {
-            unreachable!("handoff test writer only receives display transitions")
-        }
-
-        fn list_messages(&self, _query: &MessageQuery) -> Result<Vec<Message>, AtmError> {
-            unreachable!("handoff test writer only receives display transitions")
-        }
-
-        fn delete_message(&self, _key: &MessageKey) -> Result<(), AtmError> {
-            unreachable!("handoff test writer only receives display transitions")
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl AsyncMessageStore for RecordingWriter {
-        async fn apply_read_display_state_async(
-            &self,
-            _scope: MailboxScope,
-            message_ids: Vec<MessageKey>,
-            _seen_watermark: Option<IsoTimestamp>,
-        ) -> Result<(), AtmError> {
-            self.attempts
-                .fetch_add(1, std::sync::atomic::Ordering::AcqRel);
-            if self
-                .failures_remaining
-                .fetch_update(
-                    std::sync::atomic::Ordering::AcqRel,
-                    std::sync::atomic::Ordering::Acquire,
-                    |remaining| remaining.checked_sub(1),
-                )
-                .is_ok()
-            {
-                return Err(AtmError::daemon_unavailable("transient writer failure"));
-            }
-            self.applied_transitions
-                .lock()
-                .expect("recording writer mutex")
-                .push(message_ids);
-            Ok(())
-        }
     }
 
     #[test]
